requirements.txt
asgiref==3.6.0
boto3==1.26.125
botocore==1.29.125
Django==4.2
django-filter==23.2
django-storages==1.13.2
jmespath==1.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
whitenoise==6.4.0
